+++ gatsby/pages/blog/index.ts
@@ -1,7 +1,12 @@
-import { CreateBlogPages } from "./types";
-import { getPosts } from "./helpers";
-
 import { createSingleBlogPost } from "./singlePostPage";
+import { getPosts } from "./helpers";
+import { CreatePageFn, CreateRedirectFn } from "../../types";
+
+export interface CreateBlogPages {
+  graphql: Function;
+  createPage: CreatePageFn;
+  createRedirect: CreateRedirectFn;
+}
 
 export const createBlogPages = async ({
   graphql,
