+++ gatsby/pages/blog/types.ts
@@ -1,11 +1,3 @@
-import { CreatePageFn, CreateRedirectFn } from "../../types";
-
-export interface CreateBlogPages {
-  graphql: Function;
-  createPage: CreatePageFn;
-  createRedirect: CreateRedirectFn;
-}
-
 export interface PostQL {
   fields: {
     slug: string;
