+++ gatsby/pages/roadmap/index.ts
@@ -1,8 +1,13 @@
-import { CreateRoadmapPages } from "./types";
-import { getCapabilities } from "./helpers";
-
 import { createRootPage } from "./rootPage";
 import { createModalPage } from "./modalPage";
+import { getCapabilities } from "./helpers";
+import { CreatePageFn, CreateRedirectFn } from "../../types";
+
+export interface CreateRoadmapPages {
+  graphql: Function;
+  createPage: CreatePageFn;
+  createRedirect: CreateRedirectFn;
+}
 
 export const createRoadmapPages = async ({
   graphql,
